pyproject.toml
[tool.poetry]
name = "travel_task"
version = "0.1.0"
description = "Technical Task with django and django rest framework"
authors = ["Jafar Huseynov <cefer.h2018@mail.ru>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
django = ">=4.0.4"
djangorestframework = ">=3.13.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
